✨ feat(sa):

+++ 11_动态规划/dp优化/辅助数据结构dp/单调队列优化dp/二维滑动窗口最小值.go
@@ -108,7 +108,7 @@ type V = int
 // 单调队列队头元素为当前窗口最小值，队尾元素为当前窗口最大值.
 type MonoQueue struct {
 	MinQueue       []V
-	_minQueueCount []int
+	_minQueueCount []int32
 	_less          func(a, b V) bool
 	_len           int
 }
@@ -120,7 +120,7 @@ func NewMonoQueue(less func(a, b V) bool) *MonoQueue {
 }
 
 func (q *MonoQueue) Append(value V) *MonoQueue {
-	count := 1
+	count := int32(1)
 	for len(q.MinQueue) > 0 && q._less(value, q.MinQueue[len(q.MinQueue)-1]) {
 		q.MinQueue = q.MinQueue[:len(q.MinQueue)-1]
 		count += q._minQueueCount[len(q._minQueueCount)-1]
@@ -163,7 +163,7 @@ func (q *MonoQueue) String() string {
 
 type pair struct {
 	value V
-	count int
+	count int32
 }
 
 func (p pair) String() string {
